restructure each set of widget apis into a tauri plugin
